requirements.txt
gradio>=4.0.0
pandas>=2.0.0
pyyaml>=6.0
openai>=1.0.0
jinja2>=3.0.0
requests>=2.28.0

# Deep Research Dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
numpy>=1.24.0
pathlib2>=2.3.0